+++ src/main/java/com/example/insurance/dto/NewMedicineRequest.java
@@ -2,7 +2,7 @@ package com.example.insurance.dto;
 
 public record NewMedicineRequest (
         String medicineName,
-        Integer medicinePrice,
+        Double medicinePrice,
         Integer quantity
 
 ) {
